Meal specifics now load from db

+++ app/src/main/java/com/example/foodieplanner/Meal.kt
@@ -1,7 +1,9 @@
 package com.example.foodieplanner
 
-data class Meal(var name: String,
-                var ingredients: ArrayList<Ingredient>,
-                var instructions: ArrayList<String>,
-                var albumName: String,
-                var rating: Double)
+data class Meal(
+    var name: String,
+    var ingredients: ArrayList<Ingredient>,
+    var instructions: ArrayList<String>,
+    var albumName: String,
+    var rating: Float
+)
